requirements.txt
coverage==4.4.1
jsonschema==2.6.0
nose==1.3.7
six==1.11.0
PyYAML==3.12
